+++ tsetmc/instruments.py
@@ -836,7 +836,7 @@ class Instrument:
                     'holder': _pl.String,
                     'shares/units': _pl.Int64,
                     '%': _pl.Float64,
-                    'change': _pl.Float64,
+                    'change': _pl.Int64,
                     'id_cisin': _pl.String,
                 }
             )
